Move BidHub to Hubs folder
